+++ java/demo/src/main/java/com/google/i18n/phonenumbers/demo/InputForm.java
@@ -16,7 +16,7 @@
  * @author Shaopeng Jia
  */
 
-package example;
+package com.google.i18n.phonenumbers.demo;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
