+++ java/demo/src/main/webapp/WEB-INF/web.xml
@@ -1,21 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns="http://java.sun.com/xml/ns/javaee"
-  xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
-  xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-  http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" version="2.5">
+    version="2.5"
+    xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
+  http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
   <servlet>
+    <servlet-class>com.google.i18n.phonenumbers.demo.Results</servlet-class>
     <servlet-name>PhoneNumberParser</servlet-name>
-    <servlet-class>example.Results</servlet-class>
+  </servlet>
+  <servlet>
+    <servlet-class>com.google.i18n.phonenumbers.demo.InputForm</servlet-class>
+    <servlet-name>InputForm</servlet-name>
   </servlet>
   <servlet-mapping>
     <servlet-name>PhoneNumberParser</servlet-name>
     <url-pattern>/phonenumberparser</url-pattern>
   </servlet-mapping>
-  <servlet>
-    <servlet-name>InputForm</servlet-name>
-    <servlet-class>example.InputForm</servlet-class>
-  </servlet>
   <servlet-mapping>
     <servlet-name>InputForm</servlet-name>
     <url-pattern>/inputform</url-pattern>
